% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LA_Gui.R
\name{autoavailable}
\alias{autoavailable}
\title{autoavailable}
\usage{
autoavailable(column = "")
}
\arguments{
\item{column}{The column to return}
}
\description{
Function that returns unique values of a column
}
