% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LA_Gui.R
\name{autoavailableSpecies}
\alias{autoavailableSpecies}
\title{autoavailableSpecies}
\usage{
autoavailableSpecies(ro = "common")
}
\arguments{
\item{return}{order ro, common or scientific}
}
\description{
Function that help autopopulate Species in the html form
}
