% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LA_Gui.R
\name{autoavailableStatdistrict}
\alias{autoavailableStatdistrict}
\title{auto_availableStadistrict}
\usage{
autoavailableStatdistrict()
}
\description{
Function that help autopopulate District in the html form
}
