% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LA_Gui.R
\name{r.read}
\alias{r.read}
\title{r.read}
\usage{
r.read(
  proj,
  years,
  firstnames,
  lastnames,
  lfas,
  districts,
  sdistricts,
  communities,
  portcodes,
  codeports,
  provinces,
  docname,
  abstractname,
  pagesname,
  speciesnames,
  speciescodes,
  Ad,
  Ar,
  As,
  Ba,
  By,
  Ca,
  Cs,
  Ct,
  Cl,
  Co,
  Cr,
  Cu,
  De,
  Dv,
  Dr,
  Ef,
  En,
  Fr,
  Ge,
  Hi,
  Im,
  In,
  Id,
  It,
  Jo,
  La,
  Le,
  Lo,
  Ma,
  Mt,
  Mi,
  Mo,
  Ms,
  Ne,
  Of,
  Fi,
  Pa,
  Po,
  Pr,
  Pc,
  Ra,
  Re,
  Se,
  Sl,
  So,
  Su,
  Sc,
  Te,
  Ts,
  Tr,
  Ta,
  Up,
  Vi,
  Vn,
  Vo,
  Wo,
  strict
)
}
\value{
list of uri's to webpage
}
\description{
Function that read archived data and returns files
}
